apps/social-monitor/requirements.txt
# Core
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Mastodon
Mastodon.py==1.8.1

# HTTP Client
httpx==0.25.2

# Utilities
python-dateutil==2.8.2


